+++ src/components/layout/layout.tsx
@@ -100,51 +100,3 @@ export default function Layout({favouriteCount} : LayoutProps): JSX.Element | nu
     </div>
   );
 }
-
-
-// //   return (
-// //     <div className={`page${pageClassName}`}>
-// //       <header className="header">
-// //         <div className="container">
-// //           <div className="header__wrapper">
-// //             <div className="header__left">
-// //               {shouldRenderLogoMainPage && <LogoMainPage />}
-// //               {!shouldRenderLogoMainPage && <Logo/>}
-// //             </div>
-// //             {
-// //               shouldRenderUser && (
-// //                 <nav className="header__nav">
-// //                   <ul className="header__nav-list">
-// //                     <li className="header__nav-item user">
-// //                       <Link className="header__nav-link header__nav-link--profile" to={AppRoute.Favorites}>
-// //                         <div className="header__avatar-wrapper user__avatar-wrapper">
-// //                         </div>
-// //                         {isAuth &&
-// //                           <>
-// //                             <span className="header__user-name user__name">{email}</span>
-// //                             <span className="header__favorite-count">{favouriteCount}</span>
-// //                           </>}
-// //                       </Link>
-// //                       {!isAuth &&
-// //                         <Link className="header__nav-link header__nav-link--profile" to={AppRoute.Login}>
-// //                           <span className="header__login">Sign in</span>
-// //                         </Link>}
-// //                     </li>
-// //                     {isAuth &&
-// //                       <li className="header__nav-item">
-// //                         <Link className="header__nav-link" to={AppRoute.Root}>
-// //                           <span className="header__signout">Sign out</span>
-// //                         </Link>
-// //                       </li>}
-// //                   </ul>
-// //                 </nav>
-// //               )
-// //             }
-// //           </div>
-// //         </div>
-// //       </header>
-// //       <Outlet />
-// //       {shouldRenderFooter && <Footer />}
-// //     </div>
-// //   );
-// // }
