+++ src/components/login-form.tsx
@@ -55,7 +55,3 @@ export default function LoginForm(): JSX.Element {
     </form>
   );
 }
-
-
-// Валидация формы - хуком
-// ///////////////////////////////////////////////////////////
