+++ src/components/private-route.tsx
@@ -1,19 +1,14 @@
 import {Navigate, useLocation} from 'react-router-dom';
-import { AppRoute } from '../const';
 import { ReactNode } from 'react';
+
+import { AppRoute } from '../const';
 import { useAppSelector } from '../hooks/store-hooks';
 import { selectUserInfo } from '../store/slices/user-slice';
 
-
 type ProtectedRouteProps = {
-  // children: JSX.Element;
   children: ReactNode;
   onlyUnauth?: boolean;
 };
-
-// type FromState = {
-//   from?: Location;
-// };
 
 interface LocationState {
   from?: {
@@ -30,7 +25,6 @@ export default function ProtectedRoute({ children, onlyUnauth }: ProtectedRouteP
     const from = (location.state as LocationState)?.from?.pathname || AppRoute.Root;
     return <Navigate to = {from} replace />;
   }
-  //нас надо переадресовать на страницу главное или на ту страницу, на которой мы находились, когда нас редиректнуло
 
 
   if (!onlyUnauth && !user) {
@@ -40,18 +34,3 @@ export default function ProtectedRoute({ children, onlyUnauth }: ProtectedRouteP
 
   return children;
 }
-
-
-// interface PrivateRouteProps {
-//   authorizationStatus: AuthorizationStatus;
-//   isReverse?: boolean;
-//   children: JSX.Element;
-// }
-
-// export default function PrivateRoute({authorizationStatus, isReverse, children}: PrivateRouteProps): JSX.Element {
-//   return (
-//     authorizationStatus === (isReverse ? AuthorizationStatus.NoAuth : AuthorizationStatus.Auth)
-//       ? children
-//       : <Navigate to={isReverse ? AppRoute.Root : AppRoute.Login} />
-//   );
-// }
