--- src/types/favorite.ts
@@ -1,65 +1,0 @@
-import { CityName } from '../const';
-
-export type FavoriteOffer =
-{
-  id: string;
-  title: string;
-  type: string;
-  price: number;
-  city: {
-    name: CityName;
-    location: {
-    latitude: number;
-    longitude: number;
-    zoom: number;
-    };
-  };
-  location: {
-    latitude: number;
-    longitude: number;
-    zoom: number;
-  };
-  isFavorite: boolean;
-  isPremium: boolean;
-  rating: number;
-  previewImage: string;
-  description: string;
-  bedrooms: number;
-  goods: [string];
-  host: {
-    name: string;
-    avatarUrl: string;
-    isPro: boolean;
-  };
-  images: [string];
-  maxAdults: number;
-}
-
-export type FavoriteOffers = FavoriteOffer[];
-
-////////////////////////////////////////////////////
-// export type FavoriteOffer = {
-//   id: string;
-//   title: string;
-//   type: string;
-//   price: number;
-//   city: {
-//     name: CityName;
-//     location: {
-//       latitude: number;
-//       longitude: number;
-//       zoom: number;
-//     };
-//   };
-//   location: {
-//     latitude: number;
-//     longitude: number;
-//     zoom: number;
-//   };
-//   isFavorite: boolean;
-//   isPremium: boolean;
-//   rating: number;
-//   previewImage: string;
-// }
-
-// export type FavoriteOffers = FavoriteOffer[];
